requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
dnspython==1.16.0
future==0.18.2
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
protobuf==3.6.1
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-telegram-bot==12.3.0
requests==2.22.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
